requirements.txt
# ASML Valuation Framework — Python Dependencies
# Pinned versions for reproducibility

pandas==2.2.3
numpy==2.1.3
matplotlib==3.9.3
scipy==1.14.1
yfinance==0.2.51
pyyaml==6.0.2
requests==2.32.3
